tests/wpt/tests/css/css-borders/tentative/parsing/box-shadow-offset-invalid.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>CSS Borders and Box Decorations 4 Test: Parsing 'box-shadow-offset' with invalid values</title>
<link rel="author" title="Sebastian Zartner" href="mailto:sebastianzartner@gmail.com">
<link rel="help" href="https://drafts.csswg.org/css-borders-4/#box-shadow-offset">
<meta name="assert" content="This test checks that 'box-shadow-offset' supports only properly defined offset values.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/parsing-testcommon.js"></script>

<script>
test_invalid_value("box-shadow-offset", "auto");
test_invalid_value("box-shadow-offset", "1 2");
test_invalid_value("box-shadow-offset", "1% 2%");
test_invalid_value("box-shadow-offset", "1px 2px 3px");
</script>
